# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Reward Loyalty API",
        "description": "API for the Reward Loyalty script.",
        "contact": {
            "email": "support@nowsquare.com"
        },
        "version": "1.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Local API Server"
        },
        {
            "url": "https://reward-loyalty-demo.nowsquare.com/api",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/{locale}/v1/partner/cards/{cardUID}/{memberUID}/transactions/purchases": {
            "post": {
                "tags": [
                    "Partner",
                    "Transaction",
                    "Card",
                    "Staff"
                ],
                "summary": "Add a new purchase to a card",
                "description": "Endpoint for a partner to add a new purchase to a card that they own",
                "operationId": "addPurchase",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardUID",
                        "in": "path",
                        "description": "Unique identifier of the card to which the transaction will be added",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    },
                    {
                        "name": "memberUID",
                        "in": "path",
                        "description": "Unique identifier for the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Purchase data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "purchase_amount": {
                                        "description": "The amount of money spent on the purchase",
                                        "type": "number",
                                        "example": 100
                                    },
                                    "image": {
                                        "description": "The image associated with the transaction",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "description": "An optional note for the purchase",
                                        "type": "string",
                                        "example": "",
                                        "nullable": true
                                    },
                                    "staffId": {
                                        "description": "Staff ID",
                                        "type": "string",
                                        "format": "Kra8\\Snowflake\\HasSnowflakePrimary",
                                        "example": "50510833641460081"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Purchase created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        },
        "/{locale}/v1/partner/cards/{cardUID}/{memberUID}/transactions/points": {
            "post": {
                "tags": [
                    "Partner",
                    "Transaction",
                    "Card",
                    "Staff"
                ],
                "summary": "Add new points to a card",
                "description": "Endpoint for a partner to add new points to a card that they own",
                "operationId": "addPoints",
                "parameters": [
                    {
                        "name": "locale",
                        "in": "path",
                        "description": "Locale setting (e.g., `en-us`)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "default": "en-us"
                        }
                    },
                    {
                        "name": "cardUID",
                        "in": "path",
                        "description": "Unique identifier of the card to which the transaction will be added",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    },
                    {
                        "name": "memberUID",
                        "in": "path",
                        "description": "Unique identifier for the member",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "xxx-xxx-xxx-xxx",
                            "example": "123-456-789-012"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Points data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "points": {
                                        "description": "The number of points to add",
                                        "type": "number",
                                        "example": 100
                                    },
                                    "image": {
                                        "description": "The image associated with the transaction",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "note": {
                                        "description": "An optional note for the points addition",
                                        "type": "string",
                                        "example": "",
                                        "nullable": true
                                    },
                                    "staffId": {
                                        "description": "Staff ID",
                                        "type": "string",
                                        "format": "Kra8\\Snowflake\\HasSnowflakePrimary",
                                        "example": "50510833641460081"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Points added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transaction"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized access",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthenticatedResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Card not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "partner_auth_token": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "UnauthenticatedResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "Unauthenticated."
                    }
                },
                "type": "object"
            },
            "NotFoundResponse": {
                "required": [
                    "message"
                ],
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "Not found."
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "message": {
                        "description": "The error message",
                        "type": "string",
                        "example": "The email must be a valid email address."
                    },
                    "errors": {
                        "description": "The validation errors",
                        "properties": {
                            "email": {
                                "type": "array",
                                "items": {
                                    "type": "string"
                                },
                                "example": [
                                    "The email must be a valid email address."
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Transaction": {
                "properties": {
                    "staff_id": {
                        "description": "The unique identifier of the staff member who created the transaction",
                        "type": "string",
                        "example": "50510833641460081"
                    },
                    "member_id": {
                        "description": "The unique identifier of the member who made the purchase",
                        "type": "string",
                        "example": "50510834128001805"
                    },
                    "card_id": {
                        "description": "The unique identifier of the card that was used for the purchase",
                        "type": "string",
                        "example": "50510826515340561"
                    },
                    "partner_name": {
                        "description": "The name of the partner",
                        "type": "string",
                        "example": "Partner Name"
                    },
                    "partner_email": {
                        "description": "The email of the partner",
                        "type": "string",
                        "example": "partner@example.com"
                    },
                    "purchase_amount": {
                        "description": "The amount of money that was spent on the purchase",
                        "type": "number"
                    },
                    "note": {
                        "description": "An optional note that can be added to the purchase",
                        "type": "string",
                        "nullable": true
                    },
                    "points_issued": {
                        "description": "The number of points issued for the purchase",
                        "type": "integer",
                        "format": "int64"
                    },
                    "reward_redeemed": {
                        "description": "Indicates whether a reward was redeemed for the purchase",
                        "type": "boolean"
                    },
                    "reward_id": {
                        "description": "The unique identifier of the reward that was redeemed",
                        "type": "string",
                        "example": "50510834538491954"
                    },
                    "transaction_date": {
                        "description": "The timestamp of the transaction",
                        "type": "string",
                        "format": "date-time"
                    },
                    "created_at": {
                        "description": "The timestamp of when the transaction was created",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "partner_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            },
            "member_auth_token": {
                "type": "http",
                "description": "Bearer {your_auth_token_here}",
                "name": "Authorization",
                "scheme": "bearer"
            }
        }
    }
}